EUI-7927 Hearings Actuals planned to actual transformation
