+++ src/hearings/containers/hearing-actuals/hearing-actuals-add-edit-summary/hearing-actuals-add-edit-summary.component.html
@@ -106,7 +106,7 @@
                     </dt>
                     <dd class="govuk-summary-list__value"></dd>
                     <dd class="govuk-summary-list__actions">
-                      <a href="javascript:void(0)" (click)="confirmActualHearingTimeAndParties(actualHearingDay)"
+                      <a href="javascript:void(0)" (click)="confirmActualHearingTimeForDay(actualHearingDay)"
                         id="hearing-timing-result-confirm-link" class="govuk-link">
                         Confirm
                       </a>&nbsp;
@@ -143,7 +143,7 @@
                     </dt>
                     <dd class="govuk-summary-list__value"></dd>
                     <dd class="govuk-summary-list__actions">
-                      <a href="javascript:void(0)" (click)="confirmActualHearingTimeAndParties(actualHearingDay)"
+                      <a href="javascript:void(0)" (click)="confirmActualPartiesForDay(actualHearingDay)"
                         id="hearing-parties-result-confirm-link" class="govuk-link">Confirm</a>&nbsp;
                       <a *ngIf="!isPaperHearing" href="javascript:void(0)"
                         [routerLink]="['/', 'hearings', 'actuals', id, 'actuals-parties', actualHearingDay.hearingDate]"
