+++ src/hearings/containers/hearing-actuals/hearing-actuals-edit-summary/hearing-actuals-edit-summary.component.ts
@@ -3,14 +3,14 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { HearingsService } from '../../../services/hearings.service';
 import * as fromHearingStore from '../../../store';
-import { HearingActualSummaryBaseComponent } from '../hearing-actual-summary-base/hearing-actual-summary-base.component';
+import { HearingActualsSummaryBaseComponent } from '../hearing-actuals-summary-base/hearing-actuals-summary-base.component';
 
 @Component({
-  selector: 'exui-hearing-actual-summary',
-  templateUrl: './hearing-actual-summary.component.html',
-  styleUrls: ['./hearing-actual-summary.component.scss']
+  selector: 'exui-hearing-actuals-edit-summary',
+  templateUrl: './hearing-actuals-edit-summary.component.html',
+  styleUrls: ['./hearing-actuals-edit-summary.component.scss']
 })
-export class HearingActualSummaryComponent extends HearingActualSummaryBaseComponent {
+export class HearingActualsEditSummaryComponent extends HearingActualsSummaryBaseComponent {
 
   constructor(
     public readonly hearingStore: Store<fromHearingStore.State>,
@@ -27,7 +27,6 @@ export class HearingActualSummaryComponent extends HearingActualSummaryBaseCompo
   }
 
   public onSubmitHearingDetails(): void {
-    this.submitted = true;
     this.hearingStore.dispatch(new fromHearingStore.SubmitHearingActuals(this.id));
   }
 }
