+++ src/hearings/containers/hearing-actuals/index.ts
@@ -1,5 +1,5 @@
-import { HearingActualAddEditSummaryComponent } from './hearing-actual-add-edit-summary/hearing-actual-add-edit-summary.component';
-import { HearingActualSummaryComponent } from './hearing-actual-edit-summary/hearing-actual-summary.component';
+import { HearingActualsAddEditSummaryComponent } from './hearing-actuals-add-edit-summary/hearing-actuals-add-edit-summary.component';
+import { HearingActualsEditSummaryComponent } from './hearing-actuals-edit-summary/hearing-actuals-edit-summary.component';
 import { HearingActualsFinalConfirmationComponent } from './hearing-actuals-final-confirmation/hearing-actuals-final-confirmation.component';
 import { HearingActualsTimingComponent } from './hearing-actuals-timing/hearing-actuals-timing.component';
 import { HearingActualsViewEditPartiesComponent } from './hearing-actuals-view-edit-parties/hearing-actuals-view-edit-parties.component';
@@ -9,8 +9,8 @@ import { HearingStageResultComponent } from './hearing-stage-result/hearing-stag
 export const containers: any[] = [
   HearingActualsComponent,
   HearingActualsTimingComponent,
-  HearingActualAddEditSummaryComponent,
-  HearingActualSummaryComponent,
+  HearingActualsAddEditSummaryComponent,
+  HearingActualsEditSummaryComponent,
   HearingStageResultComponent,
   HearingActualsFinalConfirmationComponent,
   HearingActualsViewEditPartiesComponent,
